+++ erpnext/stock/doctype/stock_reservation_entry/stock_reservation_entry.py
@@ -1587,7 +1587,7 @@ def create_stock_reservation_entries_for_so_items(
 ):
 	"""Creates Stock Reservation Entries for Sales Order Items."""
 
-	from erpnext.selling.doctype.sales_order.sales_order import get_unreserved_qty
+	from erpnext.selling.doctype.sales_order.services.stock_reservation import get_unreserved_qty
 
 	if not from_voucher_type and (
 		sales_order.get("_action") == "submit"
